fix: static int OpenDXA::PTM::toPtmStructureType(OpenDXA::StructureType): Assertion `false' failed

+++ bindings/python/opendxa_py/include/opendxa_py/wrappers/dislocation_analysis.hpp
@@ -46,8 +46,8 @@ private:
     
     // Validation methods
     void validateInputFile(const std::string& filePath) const;
-    void validateCircuitSize(int size) const;
-    void validateStretchability(int stretch) const;
+    void validateCircuitSize(double size) const;
+    void validateStretchability(double stretch) const;
 };
 
 }
